Added rainbow 6 stat and op commands, modularized command files, cleaned up file structure

+++ DiscordPBot/RainbowSix/Hostage.cs
@@ -17,6 +17,6 @@ namespace DiscordPBot.RainbowSix
         [JsonProperty("wins")]
         public int Wins { get; set; }
         [JsonProperty("wl")]
-        public int Wl { get; set; }
+        public float Wl { get; set; }
     }
 }
